refactor receipt into sub components
